Bug 1309109 part 1 - Change underlying type of StyleBackendType to uint8_t. r=heycam
MozReview-Commit-ID: B6TmOq6elfs

UltraBlame original commit: dc64dced730033eb9bea3d4783cbae6bb5a8e874

+++ layout/style/StyleBackendType.h
@@ -12,7 +12,7 @@ namespace mozilla {
 
 
 
-enum class StyleBackendType : int
+enum class StyleBackendType : uint8_t
 {
   Gecko = 1,
   Servo
